refactor: useGameStatus로 커스텀 훅 분리
